Rename cloudconfig templates into ignition
