+++ service/controller/v5/templates/ignition/small.go
@@ -1,4 +1,4 @@
-package cloudconfig
+package ignition
 
 const Small = `---
 ignition:
